Update R packages
These updates were done on Ubuntu 16.04 with Nix 2.1.3 installed by doing

```bash
cd pkgs/development/r-modules/
./generate.sh
```

Most of the diffs in these files are due to the list of `depends` now being arranged alphabetically. I don't know why this change occurred but can only think it must be because the line `depends <- paste(sort(unique(depends)), collapse=" ")` was added to `generate-r-packages.R` after these files were generated (?). Committing this change now so that in the future, diffs are only related to changes in the version, sha256 etc of packages.